Revert "Move to `key` module."
This reverts commit a73e3e46417eed15a75a85417ba51b8e6167275d.
