+++ src/cose.rs
@@ -1,4 +1,3 @@
-pub mod key;
 pub mod mac0;
 pub mod sign1;
