+++ src/definitions/device_key/cose_key.rs
