+++ src/definitions/device_key/mod.rs
@@ -3,7 +3,8 @@ use serde::{Deserialize, Serialize};
 use serde_cbor::Value as CborValue;
 use std::collections::BTreeMap;
 
-pub use crate::cose::key::CoseKey;
+pub mod cose_key;
+pub use cose_key::CoseKey;
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
