+++ src/presentation/mod.rs
@@ -25,8 +25,7 @@ impl Stringify for device::SessionManagerEngaged {}
 impl Stringify for device::SessionManager {}
 impl Stringify for reader::SessionManager {}
 
-use crate::cose::key::CoseKey;
-use crate::definitions::helpers::Tag24;
+use crate::definitions::{device_key::cose_key::CoseKey, helpers::Tag24};
 use hkdf::Hkdf;
 use sha2::Sha256;
